break preoprty-bag-column-renderer cycles
