+++ src/assessments/common/property-bag-column-renderer-factory.tsx
@@ -1,11 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { ColumnValueBag } from 'common/types/property-bag/column-value-bag';
+import { PropertyBagColumnRendererConfig } from 'common/types/property-bag/property-bag-column-renderer-config';
 import { AssessmentInstanceRowData } from 'DetailsView/components/assessment-instance-table';
-import {
-    propertyBagColumnRenderer,
-    PropertyBagColumnRendererConfig,
-} from './property-bag-column-renderer';
+import { propertyBagColumnRenderer } from './property-bag-column-renderer';
 
 export class PropertyBagColumnRendererFactory {
     public static getRenderer<TPropertyBag extends ColumnValueBag>(
