+++ src/assessments/custom-widgets/custom-widgets-column-renderer-factory.tsx
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { ColumnValueBag } from 'common/types/property-bag/column-value-bag';
+import { PropertyBagColumnRendererConfig } from 'common/types/property-bag/property-bag-column-renderer-config';
 import { AssessmentInstanceRowData } from 'DetailsView/components/assessment-instance-table';
-import { PropertyBagColumnRendererConfig } from '../common/property-bag-column-renderer';
 import { customWidgetsColumnRenderer } from './custom-widgets-column-renderer';
 
 export class CustomWidgetsColumnRendererFactory {
