+++ src/assessments/images/test-steps/image-function.tsx
@@ -3,7 +3,7 @@
 import {
     NoValue,
     PropertyBagColumnRendererConfig,
-} from 'assessments/common/property-bag-column-renderer';
+} from 'common/types/property-bag/property-bag-column-renderer-config';
 import { ImageFunctionPropertyBag } from 'common/types/property-bag/image-function';
 import { VisualizationType } from 'common/types/visualization-type';
 import { link } from 'content/link';
