+++ src/assessments/native-widgets/test-steps/instructions.tsx
@@ -14,7 +14,7 @@ import { InstructionsAndLabelsNotes } from '../../common/instructions-and-labels
 import {
     NoValue,
     PropertyBagColumnRendererConfig,
-} from '../../common/property-bag-column-renderer';
+} from '../../../common/types/property-bag/property-bag-column-renderer-config';
 import { PropertyBagColumnRendererFactory } from '../../common/property-bag-column-renderer-factory';
 import * as Markup from '../../markup';
 import { ReportInstanceField } from '../../types/report-instance-field';
