+++ src/tests/unit/tests/assessments/common/property-bag-column-renderer.test.tsx
@@ -3,10 +3,8 @@
 import { shallow, ShallowWrapper } from 'enzyme';
 import * as React from 'react';
 
-import {
-    propertyBagColumnRenderer,
-    PropertyBagColumnRendererConfig,
-} from 'assessments/common/property-bag-column-renderer';
+import { propertyBagColumnRenderer } from 'assessments/common/property-bag-column-renderer';
+import { PropertyBagColumnRendererConfig } from 'common/types/property-bag/property-bag-column-renderer-config';
 import { ColumnValueBag } from '../../../../../common/types/property-bag/column-value-bag';
 import { AssessmentInstanceRowData } from '../../../../../DetailsView/components/assessment-instance-table';
 import { DictionaryStringTo } from '../../../../../types/common-types';
