+++ src/tests/unit/tests/assessments/report-instance-field.test.ts
@@ -3,7 +3,7 @@
 import { ReportInstanceField } from 'assessments/types/report-instance-field';
 import { BagOf } from 'common/types/property-bag/column-value-bag';
 
-import { PropertyBagColumnRendererConfig } from 'assessments/common/property-bag-column-renderer';
+import { PropertyBagColumnRendererConfig } from 'common/types/property-bag/property-bag-column-renderer-config';
 
 describe('ReportInstanceField', () => {
     type Bag = { one?: string; two?: string; attr?: BagOf<string> };
